Fix bug when colla didnt exists. Add following t user

+++ back-end/api/users/__init__.py
@@ -52,12 +52,15 @@ def add_belong(user_id):
         colla_id = request.args.get('colla_id')
         if colla_id is not None:
             colla = get_colla_ctrl(DB(db_configuration).get_database_connection()).get(colla_id)
-            if user_service.can_join(user, colla):
-                user_service.add_belonging_colla(user, colla)
-                user = user_service.get_all_info(user)
-                return make_response(jsonify(user.__dict__), 200)
+            if colla:
+                if user_service.can_join(user, colla):
+                    user_service.add_belonging_colla(user, colla)
+                    user = user_service.get_all_info(user)
+                    return make_response(jsonify(user.__dict__), 200)
+                else:
+                    abort(409)
             else:
-                abort(409)
+                abort(404)
         abort(400)
 
 
@@ -72,7 +75,30 @@ def remove_belong(user_id):
         colla_id = request.args.get('colla_id')
         if colla_id is not None:
             colla = get_colla_ctrl(DB(db_configuration).get_database_connection()).get(colla_id)
-            user_service.remove_belong(user, colla)
-            user = user_service.get_all_info(user)
-            return make_response(jsonify(user.__dict__), 200)
+            if colla:
+                user_service.remove_belong(user, colla)
+                user = user_service.get_all_info(user)
+                return make_response(jsonify(user.__dict__), 200)
+            else:
+                abort(404)
+        abort(400)
+
+
+@app.route('/users/<int:user_id>/follows', methods=['POST'])
+def add_follow(user_id):
+    db_configuration = json.loads(open("api/db/db.json").read())
+    user_ctrl = get_user_ctrl(DB(db_configuration).get_database_connection())
+    user = user_ctrl.get(user_id)
+    if user is None:
+        abort(404)
+    else:
+        colla_id = request.args.get('colla_id')
+        if colla_id is not None:
+            colla = get_colla_ctrl(DB(db_configuration).get_database_connection()).get(colla_id)
+            if colla:
+                user_service.add_following(user.id, colla_id)
+                user = user_service.get_all_info(user)
+                return make_response(jsonify(user.__dict__), 200)
+            else:
+                abort(404)
         abort(400)
